+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [0.11.0](https://github.com/Blobfolio/argyle/releases/tag/v0.11.0) - 2025-02-23
 
+### New
+
+* `Argument::Path` variant (via crate feature `try_paths`)
+
 ### Changed
 
 * Bump Rust edition to 2024